requirements.txt
absl-py==0.7.1
astor==0.7.1
audioread==2.1.7
bitarray==0.9.2
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
decorator==4.4.0
Django==2.2.1
django-rest-swagger==2.2.0
djangorestframework==3.9.4
drf-yasg==1.15.0
gast==0.2.2
grpcio==1.20.1
h5py==2.9.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
librosa==0.6.3
llvmlite==0.28.0
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mmh3==2.5.1
mock==3.0.5
numba==0.43.1
numpy==1.16.3
openapi-codec==1.3.2
pandas==0.24.2
probably==1.1.0
protobuf==3.7.1
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
requests==2.22.0
resampy==0.2.1
ruamel.yaml==0.15.96
scikit-learn==0.21.1
scipy==1.3.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
uritemplate==3.0.0
urllib3==1.25.2
Werkzeug==0.15.4
